+++ hyper/.hyper.js
@@ -142,6 +142,8 @@ module.exports = {
     // bellSoundURL: 'http://example.com/bell.mp3',
 
     // for advanced config flags please refer to https://hyper.is/#cfg
+
+    confirmQuit: true,
   },
 
   // a list of plugins to fetch and install from npm
@@ -150,7 +152,9 @@ module.exports = {
   //   `hyperpower`
   //   `@company/project`
   //   `project#1.0.1`
-  plugins: [],
+  plugins: [
+    'hyper-confirm'
+  ],
 
   // in development, you can create a directory under
   // `~/.hyper_plugins/local/` and include it here